requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
comtypes==1.1.7
cycler==0.10.0
docopt==0.6.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.6.4
Js2Py==0.70
kiwisolver==1.3.0
lazy-object-proxy==1.5.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.3
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.1
pipwin==0.5.0
playsound==1.2.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/erika/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
pyjsparser==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
simpleaudio==1.0.4
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
toml==0.10.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1

